Perl API - follow on to type change in 4eaeef7f0c.

+++ contribs/perlapi/libslurm/perl/job.c
@@ -291,7 +291,7 @@ job_info_to_hv(job_info_t *job_info, HV *hv)
 	STORE_FIELD(hv, job_info, start_time, time_t);
 	if(job_info->state_desc)
 		STORE_FIELD(hv, job_info, state_desc, charp);
-	STORE_FIELD(hv, job_info, state_reason, uint16_t);
+	STORE_FIELD(hv, job_info, state_reason, uint32_t);
 	if(job_info->std_in)
 		STORE_FIELD(hv, job_info, std_in, charp);
 	if(job_info->std_out)
@@ -425,7 +425,7 @@ hv_to_job_info(HV *hv, job_info_t *job_info)
 	FETCH_FIELD(hv, job_info, show_flags, uint16_t, TRUE);
 	FETCH_FIELD(hv, job_info, start_time, time_t, TRUE);
 	FETCH_FIELD(hv, job_info, state_desc, charp, FALSE);
-	FETCH_FIELD(hv, job_info, state_reason, uint16_t, TRUE);
+	FETCH_FIELD(hv, job_info, state_reason, uint32_t, TRUE);
 	FETCH_FIELD(hv, job_info, std_in, charp, FALSE);
 	FETCH_FIELD(hv, job_info, std_out, charp, FALSE);
 	FETCH_FIELD(hv, job_info, std_err, charp, FALSE);
